+++ packages/babel-plugin-veui/test/fixtures/veui/components/source.js
@@ -1,5 +1,6 @@
 import Button from './Button'
 import { default as Select, foo } from './Select'
-import { Form, Field, default as veui } from 'veui'
+import { Input } from 'veui'
+import { Form, Field, Icon as VeuiIcon, default as veui } from 'veui'
 import Switch from 'veui/components/Switch'
 import { default as Whatever, zig } from 'veui/components/Calendar'